+++ spec/internal/app/mailers/user_mailer.rb
@@ -1,3 +1,0 @@
-class UserMailer < ApplicationMailer
-  include Avocado::Mailer
-end