+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Convert the `Avocado::Mailer` module into a class
+
 ## [0.3.0] - 2023-07-17
 
 - Add an `Avocado::Mailer` which generates each of the signed ids