+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add controller for "passwordless" email-link sign-in
 - Add event class to log user auth events
 - Add user-facing email and password edit pages
+- Add various event logging callbacks
 
 ## [0.4.0] - 2023-07-19