+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.5.0] - 2023-07-21
+
 - Add controller for "passwordless" email-link sign-in
 - Add event class to log user auth events
 - Add user-facing email and password edit pages